Move ephemeral payment store to child package
