+++ core/payments/ondemand/ephemeral/doc.go
@@ -1,0 +1,2 @@
+// Package ephemeral contains an in-memory implementation of a CumulativePaymentStore
+package ephemeral
