+++ core/payments/ondemand/ephemeral/ephemeral_cumulative_payment_store_test.go
@@ -1,4 +1,4 @@
-package ondemand
+package ephemeral
 
 import (
 	"context"
